Implement search functionality
Summary:
PTB-12 add search functionality for browsers, windowProps and documentProps; update eslint rules; adjust all files according to eslint rules

!!! Server branch - ft-PTB-16 !!!

1. initial query returns uniq browsers where versions belong to each browser
2. browser search searchs among browser names
3. click on browser's version shows corresponding document and window properties. 2nd click on the same version cell clear the filter for a current browser version and remove green color.
Click on another version cell show document and window properties for the browser version was clicked in spite of the previous cell stays green (should be improved later)
4. search for document and window properties search among represented values.
5. I had a lot of eslint errors so updated rules to have ';' always and use singlequotes

Test Plan: search should work for all three lists

Reviewers: jan, mithun, dario

Reviewed By: dario

Differential Revision: https://phabricator.meetrics.de/D2420

+++ .eslintrc.js
@@ -23,7 +23,7 @@ module.exports = {
     "generator-star-spacing": 0,
     // allow debugger during development
     "no-debugger": process.env.NODE_ENV === "production" ? 2 : 0,
-    "quotes": [2, "double"],
+    "quotes": [2, "single"],
     "semi": ["error", "always"]
   }
 }
